Refactored Component Names for new UX Design
