+++ client/src/components/traceView.tsx
